Move GP/Kernel marker classes to kernel_helpers
ConstantKernel, VariableKernel, ConstantMatrix, VariableMatrix,
NoiseMatrix (base), GP, ConstantGP, VariableGP, GlobalVariableGP now
live in kernel_helpers.py. matrix.py re-imports them so matrix.X
keeps working for existing call sites (signals.py, optimal.py, etc.).

likelihood_metamath.py drops the matrix.* isinstance dispatch on the
marker types and uses kh.Kernel / kh.ConstantGP / kh.VariableGP
directly. matrix is still imported there for the Tier-3 cglogL helpers
(matrix.cgsolve, matrix.make_logdet_estimator) and matrix.CompoundGlobalGP
(an untested edge case).

43/43 parity tests pass.
